fix up tests and streaming
