+++ core/src/languageModels/openAI.ts
@@ -1,76 +1,37 @@
 import OpenAI from "openai";
-import { RequestOptions } from "openai/core"
-import { ChatCompletionMessage, ChatCompletionMessageParam, CompletionCreateParamsNonStreaming } from "openai/resources/chat/completions"
-import { SpanStatusCode, trace } from '@opentelemetry/api';
-import {
-  ChatMessage,
-  ChatMessageRoleEnum,
-  ExecutorResponse,
-  FunctionCall,
-  FunctionSpecification,
-  LanguageModelProgramExecutor,
-  LanguageModelProgramExecutorExecuteOptions,
-  RequestOptionsStreaming,
-} from "./index";
-import { zodToSchema } from "./zodToSchema";
-import { html } from "common-tags";
-import { ReusableStream } from "./reusableStream";
+import { ChatMessage, ExecutorResponse, FunctionSpecification, LanguageModelProgramExecutor } from ".";
+import { ChatCompletionCreateParams, ChatCompletionMessageParam, ChatCompletionTool } from "openai/resources";
+import { RequestOptions } from "openai/core";
+import { trace } from "@opentelemetry/api";
+import { RunnableFunctionWithParse, RunnableTools } from "openai/lib/RunnableFunction";
+import { zodToJsonSchema } from 'zod-to-json-schema';
+import { LanguageModelProgramExecutorExecuteOptions } from "../legacy";
 import { backOff } from "exponential-backoff";
+import { ChatCompletionToolRunnerParams } from "openai/lib/ChatCompletionRunner";
+import { withErrorCatchingSpan } from "./errorCatchingSpan";
+import { ReusableStream } from "./reusableStream";
+import { OpenAICompatibleStream } from "./LLMStream";
 
-const MAX_RETRIES = 3;
+type Config = ConstructorParameters<typeof OpenAI>[0];
+type ChatCompletionParams =
+  Partial<ChatCompletionCreateParams>
+
+type DefaultCompletionParams = ChatCompletionParams & {
+  model: ChatCompletionCreateParams["model"];
+};
 
 const tracer = trace.getTracer(
   'open-souls-openai',
   '0.0.1',
 );
 
-export enum Model {
-  GPT_4 = "gpt-4",
-  GPT_4_preview = "gpt-4-1106-preview",
-  GPT_3_5_turbo = "gpt-3.5-turbo",
-  GPT_3_5_turbo_0613 = "gpt-3.5-turbo-0613",
-  GPT_3_5_turbo_16k = "gpt-3.5-turbo-16k",
-  GPT_3_5_turbo_preview = "gpt-3.5-turbo-1106",
-}
+// eslint-disable-next-line @typescript-eslint/no-empty-function
+const emptyFunction = () => { }
 
-type Config = ConstructorParameters<typeof OpenAI>[0];
-
-type ChatCompletionParams =
-  Partial<CompletionCreateParamsNonStreaming>
-
-type DefaultCompletionParams = ChatCompletionParams & {
-  model: Model | string;
-};
-
-enum FunctionCallErrorType {
-  NoFunctionCall = "No function call was returned",
-  FunctionNotFound = "No function by that name",
-  InvalidJSON = "Invalid JSON",
-  SchemaMismatch = "Schema mismatch",
-}
-
-interface FunctionCallError {
-  type: FunctionCallErrorType;
-  description: string
-}
-
-type ValidateFunctionCallResult =
-  | { error: FunctionCallError; parsed?: never }
-  | { error?: never; parsed: any };
-
-
-interface RetryInformation {
-  functionCall?: FunctionCall
-  error: FunctionCallError;
-  retryCount: number
-}
-
-
-export class OpenAILanguageProgramProcessor
-  implements LanguageModelProgramExecutor {
+export class OpenAILanguageProgramProcessor implements LanguageModelProgramExecutor {
   client: OpenAI;
-  defaultCompletionParams: DefaultCompletionParams;
-  defaultRequestOptions: RequestOptions;
+  defaultCompletionParams: DefaultCompletionParams
+  defaultRequestOptions: RequestOptions
 
   constructor(
     openAIConfig: Config = {},
@@ -85,7 +46,7 @@ export class OpenAILanguageProgramProcessor
       ...openAIConfig,
     });
     this.defaultCompletionParams = {
-      model: Model.GPT_3_5_turbo,
+      model: "gpt-3.5-turbo",
       ...defaultCompletionParams,
       stream: false,
     };
@@ -95,347 +56,238 @@ export class OpenAILanguageProgramProcessor
     }
   }
 
-  private validateFunctioncall(requestedFunction: LanguageModelProgramExecutorExecuteOptions["functionCall"], message: ChatCompletionMessage | undefined, functions: FunctionSpecification[]): ValidateFunctionCallResult {
-    if (!requestedFunction) {
-      return {
-        parsed: message?.content
-      }
-    }
-    const functionCall = message?.function_call;
-
-    if (!functionCall) {
-      return {
-        error: {
-          type: FunctionCallErrorType.NoFunctionCall,
-          description: "No function was called",
-        }
-      }
-    }
-
-    const fn = functions.find((fn) => fn.name === functionCall.name);
-    if (!fn) {
-      return {
-        error: {
-          type: FunctionCallErrorType.FunctionNotFound,
-          description: `No function by the name ${functionCall.name}. Valid function names are: ${functions.map((fn) => fn.name).join(", ")}`,
-        }
-      };
-    }
-
-    let jsonParsed = {};
-    try {
-      jsonParsed = JSON.parse(functionCall.arguments);
-    } catch (err: any) {
-      return {
-        error: {
-          type: FunctionCallErrorType.InvalidJSON,
-          description: `Invalid JSON: ${err.message}`,
-        }
-      };
-    }
-
-    try {
-      const zodParsed = fn.parameters.safeParse(jsonParsed);
-      if (!zodParsed.success) {
-        return {
-          error: {
-            type: FunctionCallErrorType.SchemaMismatch,
-            description: `Schema mismatch: ${zodParsed.error.message}`,
-          }
-        };
-      }
-
-      return {
-        parsed: zodParsed.data
-      }
-    } catch (err: any) {
-      return {
-        error: {
-          type: FunctionCallErrorType.SchemaMismatch,
-          description: `Schema mismatch: ${err.message}`,
-        }
-      };
-    }
-  }
-
-  private streamingExecute<FunctionCallReturnType = undefined>(
-    messages: ChatMessage[],
-    completionParams: LanguageModelProgramExecutorExecuteOptions = {},
-    functions: FunctionSpecification[] = [],
-    requestOptions: RequestOptionsStreaming = { stream: true },
-    retryError: RetryInformation | undefined = undefined
-  ): Promise<{
-    response: Promise<ExecutorResponse<FunctionCallReturnType>>,
-    stream: AsyncIterable<string>,
-  }> {
-    return tracer.startActiveSpan('streamingExecute', async (span) => {
-      try {
-        const { functionCall, ...restRequestParams } = completionParams;
-
-        const params = {
-          ...this.defaultCompletionParams,
-          ...restRequestParams,
-          function_call: functionCall,
-          messages: messages as ChatCompletionMessageParam[],
-        }
-        if (functions.length > 0) {
-          params.functions = functions.map((fn) => {
-            return {
-              ...fn,
-              parameters: zodToSchema(fn.parameters),
-            };
-          });
-        }
-
-        if (retryError) {
-          if (retryError.retryCount >= MAX_RETRIES) {
-            throw new Error(`Exceeded max retries of ${MAX_RETRIES} for error: ${JSON.stringify(retryError.error)}`);
-          }
-          span.setAttributes({
-            "retry-attempt": retryError.retryCount + 1,
-            "retry-error": JSON.stringify(retryError.error),
-            "retry-function-call": JSON.stringify(retryError.functionCall),
-          })
-
-          params.messages = [
-            ...params.messages,
-            {
-              role: ChatMessageRoleEnum.Assistant,
-              content: "",
-              function_call: retryError.functionCall,
-            },
-            {
-              role: ChatMessageRoleEnum.User,
-              content: html`
-                Function call resulted in error: ${JSON.stringify(retryError.error)}
-              `
-            }
-          ]
-        }
-
-        span.setAttributes({
-          "params": JSON.stringify(params),
-          "request-options": JSON.stringify(requestOptions || {}),
-        });
-
-        const stream = await backOff(() => {
-          return this.client.chat.completions.create(
-            { ...params, stream: true },
-            {
-              ...this.defaultRequestOptions,
-              ...requestOptions,
-            }
-          )
-        }, {
-          maxDelay: 200,
-          numOfAttempts: 3,
-          retry: (e, attempt) => {
-            console.error("error in open ai call", e, attempt)
-            return true
-          }
-        })
-
-        const _generateContent = async function* () {
-          for await (const res of stream) {
-            if (res.choices && res.choices.length > 0) {
-              const message = res.choices[0].delta;
-              if (message) {
-                const data = {
-                  content: message.content,
-                  functionCall: message.function_call
-                };
-                yield data
-              }
-            }
-          }
-        }
-
-        const reusableStream = new ReusableStream(_generateContent())
-
-        const streamToText = async function* () {
-          for await (const res of reusableStream.stream()) {
-            yield res.content || res.functionCall?.arguments || "";
-          }
-        }
-
-        const respFunction = async () => {
-          let returnedFunctionallArguments = ""
-          let content = ""
-          for await (const res of reusableStream.stream()) {
-            if (res.functionCall) {
-              returnedFunctionallArguments += res.functionCall.arguments
-            } else {
-              content += res.content || ""
-            }
-          }
-
-          let parsedFnCall: { name: string, arguments: string } | undefined = undefined
-
-          if (functionCall && returnedFunctionallArguments) {
-            if (typeof functionCall === "string") {
-              console.error("string function call is unsupported")
-              throw new Error('string function call is unsupported')
-            }
-            parsedFnCall = {
-              ...functionCall,
-              arguments: returnedFunctionallArguments
-            }
-          }
-
-          span.setAttributes({
-            "completion-content": content,
-            "completion-function-call": JSON.stringify(returnedFunctionallArguments || "{}"),
-          })
-
-          const { error, parsed } = this.validateFunctioncall(functionCall, {
-            content: content,
-            function_call: parsedFnCall,
-            role: ChatMessageRoleEnum.Assistant,
-          }, functions);
-          if (error) {
-            console.error("error in execute", error)
-            throw new Error('error in execute')
-          }
-
-          return {
-            content: content,
-            functionCall: parsedFnCall as (FunctionCall | undefined),
-            parsedArguments: parsed as FunctionCallReturnType
-          };
-        }
-
-        return {
-          response: respFunction(),
-          stream: streamToText(),
-        };
-
-      } catch (err: any) {
-        span.recordException(err);
-        console.error('error in execute', err);
-        throw err;
-      } finally {
-        span.end();
-      }
-    })
-  }
-
-  async execute<FunctionCallReturnType = undefined>(
+  async execute(
     messages: ChatMessage[],
     completionParams: LanguageModelProgramExecutorExecuteOptions = {},
     functions: FunctionSpecification[] = [],
     requestOptions: RequestOptions = {},
-    retryError: RetryInformation | undefined = undefined
   ): Promise<any> {
+    return withErrorCatchingSpan(tracer, "execute", async (span) => {
+      const { functionCall, ...completionParamsWithoutFunctionCall } = completionParams;
 
-    if (requestOptions.stream) {
-      return this.streamingExecute<FunctionCallReturnType>(messages, completionParams, functions, { ...requestOptions, stream: true }, retryError)
-    }
-
-    return tracer.startActiveSpan('execute', async (span) => {
-      try {
-        const { functionCall, ...completionParamsWithoutFunctionCall } = completionParams;
-
-        const params = {
-          ...this.defaultCompletionParams,
-          ...completionParamsWithoutFunctionCall,
-          function_call: functionCall,
-          messages: messages as ChatCompletionMessageParam[],
-        }
-        if (functions.length > 0) {
-          params.functions = functions.map((fn) => {
-            return {
-              ...fn,
-              parameters: zodToSchema(fn.parameters),
-            };
-          });
-        }
-
-        if (retryError) {
-          if (retryError.retryCount >= MAX_RETRIES) {
-            throw new Error(`Exceeded max retries of ${MAX_RETRIES} for error: ${JSON.stringify(retryError.error)}`);
-          }
-          span.setAttributes({
-            "retry-attempt": retryError.retryCount + 1,
-            "retry-error": JSON.stringify(retryError.error),
-            "retry-function-call": JSON.stringify(retryError.functionCall),
-          })
-
-          params.messages = [
-            ...params.messages,
-            {
-              role: ChatMessageRoleEnum.Assistant,
-              content: "",
-              function_call: retryError.functionCall,
-            },
-            {
-              role: ChatMessageRoleEnum.User,
-              content: html`
-                Function call resulted in error: ${JSON.stringify(retryError.error)}
-              `
-            }
-          ]
-        }
-
-        span.setAttributes({
-          "params": JSON.stringify(params),
-          "request-options": JSON.stringify(requestOptions || {}),
-        });
-
-        const res = await backOff(() => {
-          return this.client.chat.completions.create(
-            { ...params, stream: false },
-            {
-              ...this.defaultRequestOptions,
-              ...requestOptions,
-            }
-          )
-        }, {
-          maxDelay: 200,
-          numOfAttempts: 3,
-          retry: (e, attempt) => {
-            console.error("error in open ai call", e, attempt)
-            return true
-          }
-        })
-
-        span.setAttributes({
-          "total-tokens": res.usage?.total_tokens || "?",
-          "prompt-tokens": res.usage?.prompt_tokens || "?",
-          "completion-tokens": res.usage?.completion_tokens || "?",
-          "completion-content": res?.choices[0]?.message?.content || "?",
-          "completion-function-call": JSON.stringify(res?.choices[0]?.message?.function_call || "{}"),
-        })
-
-        const { error, parsed } = this.validateFunctioncall(functionCall, res?.choices[0]?.message, functions);
-        if (error) {
-          if (retryError?.retryCount || 0 >= MAX_RETRIES) {
-            throw new Error(`Exceeded max retries of ${MAX_RETRIES} for error: ${JSON.stringify(error)}`);
-          }
-          span.setStatus({
-            code: SpanStatusCode.ERROR,
-            message: JSON.stringify(error),
-          });
-          return this.execute(messages, completionParams, functions, requestOptions, {
-            error: error,
-            retryCount: retryError?.retryCount ?? 1,
-            functionCall: res?.choices[0]?.message?.function_call,
-          });
-        }
-
-        return {
-          content: res?.choices[0]?.message?.content,
-          functionCall: res?.choices[0]?.message?.function_call,
-          parsedArguments: parsed as FunctionCallReturnType
-        };
-      } catch (err: any) {
-        span.recordException(err);
-        console.error('error in execute', err);
-        throw err;
-      } finally {
-        span.end();
+      const params: ChatCompletionCreateParams = {
+        ...this.defaultCompletionParams,
+        ...completionParamsWithoutFunctionCall,
+        messages: messages as ChatCompletionMessageParam[],
       }
 
-    })
+      span.setAttributes({
+        "params": JSON.stringify(params),
+        "request-options": JSON.stringify(requestOptions || {}),
+      });
 
+      if (requestOptions.stream) {
+        if (functionCall) {
+          return this.executeStreaming({
+            ...params,
+            tool_choice: {
+              type: "function",
+              function: functionCall as { name: string }
+            },
+          }, requestOptions, functions)
+        }
+        return this.executeStreaming(params, requestOptions, functions)
+      }
+
+      if (functionCall) {
+        return this.executeWithFunctionCall(params, requestOptions, (functionCall as any), functions)
+      }
+
+      return this.nonFunctionExecute(params, requestOptions)
+    })
+  }
+
+  private async executeStreaming(
+    completionParams: ChatCompletionCreateParams,
+    requestOptions: RequestOptions,
+    functions: FunctionSpecification[] = [],
+  ): Promise<{
+    response: Promise<ExecutorResponse<any>>,
+    stream: AsyncIterable<string>,
+  }> {
+    return tracer.startActiveSpan("executeStreaming", async (span) => {
+      try {
+        const tools = functions.length > 0 ? this.mapFunctionCallsToTools(functions, emptyFunction) as unknown as ChatCompletionTool[] : undefined
+
+        const params = {
+          ...completionParams,
+          ...(tools ? { tools } : {})
+        }
+  
+        const rawStream = await this.client.chat.completions.create(
+          { ...params, stream: true },
+          {
+            ...this.defaultRequestOptions,
+            ...requestOptions,
+          }
+        )
+
+        const stream = new OpenAICompatibleStream(rawStream)
+  
+        const streamToText = async function* () {
+          for await (const res of stream.stream()) {
+            yield res.choices[0].delta.content || res.choices[0].delta.tool_calls?.[0]?.function?.arguments || ""
+          }
+        }
+  
+        const responseFn = async () => {
+          const content = (await stream.finalContent()) || ""
+          const functionCall = await (stream.finalFunctionCall())
+          let parsed: any = undefined
+  
+          if (functionCall) {
+            const fn = functions.find((f) => f.name === functionCall.name)
+            parsed = fn?.parameters.parse(JSON.parse(functionCall.arguments))
+          }
+    
+          span.setAttributes({
+            "completion-content": content,
+            "completion-function-call": JSON.stringify(functionCall || "{}"),
+            "completion-parsed": JSON.stringify(parsed || "{}"),
+          })
+
+          span.end()
+  
+          return {
+            content,
+            functionCall: parsed ? { name: functionCall?.name || "", arguments: parsed } : { name: "", arguments: "" },
+            parsedArguments: parsed ? parsed : content,
+          };
+        }
+  
+        return {
+          response: responseFn(),
+          stream: streamToText(),
+        }
+      } catch (err: any) {
+        span.recordException(err);
+        console.error("error in open ai call", err)
+        span.end()
+        throw err
+      }
+    })
+  }
+
+  private async executeWithFunctionCall(
+    completionParams: ChatCompletionCreateParams,
+    requestOptions: RequestOptions,
+    functionCall: { name: string },
+    functions: FunctionSpecification[] = [],
+  ): Promise<any> {
+    return withErrorCatchingSpan(tracer, "executeWithFunctionCall", async (span) => {
+      let parsed: any
+
+      const handler = (fnCall: any) => {
+        parsed = fnCall
+      }
+
+      const paramsWithFunctions: ChatCompletionToolRunnerParams<any> = {
+        ...completionParams,
+        tool_choice: {
+          type: "function",
+          function: functionCall
+        },
+        tools: this.mapFunctionCallsToTools(functions, handler),
+        stream: false,
+      }
+
+      if (!paramsWithFunctions.model) {
+        throw new Error("missing model")
+      }
+
+      const { content, usage } = await backOff(async () => {
+        const runner = this.client.beta.chat.completions.runTools(
+          { ...paramsWithFunctions },
+          {
+            ...this.defaultRequestOptions,
+            ...requestOptions,
+          }
+        )
+        runner.on("error", (error) => console.error("runner error", error))
+        return {
+          content: await runner.finalContent(),
+          usage: await runner.totalUsage(),
+        }
+      }, {
+        maxDelay: 200,
+        numOfAttempts: 3,
+        retry: (e, attempt) => {
+          console.error("error in open ai call", e, attempt)
+          return true
+        }
+      })
+
+      span.setAttributes({
+        "total-tokens": usage.total_tokens || "?",
+        "prompt-tokens": usage.prompt_tokens || "?",
+        "completion-tokens": usage.completion_tokens || "?",
+        "completion-content": "",
+        "completion-function-call": JSON.stringify(content),
+        "completion-parsed": JSON.stringify(parsed),
+      })
+
+      return {
+        content: "",
+        functionCall: content,
+        parsedArguments: parsed
+      };
+    })
+  }
+
+  private async nonFunctionExecute(
+    completionParams: ChatCompletionCreateParams,
+    requestOptions: RequestOptions,
+  ): Promise<any> {
+    return withErrorCatchingSpan(tracer, "nonFunctionExecute", async (span) => {
+      const res = await backOff(() => {
+        return this.client.chat.completions.create(
+          { ...completionParams, stream: false },
+          {
+            ...this.defaultRequestOptions,
+            ...requestOptions,
+          }
+        )
+      }, {
+        maxDelay: 200,
+        numOfAttempts: 3,
+        retry: (e, attempt) => {
+          console.error("error in open ai call", e, attempt)
+          return true
+        }
+      })
+
+      span.setAttributes({
+        "total-tokens": res.usage?.total_tokens || "?",
+        "prompt-tokens": res.usage?.prompt_tokens || "?",
+        "completion-tokens": res.usage?.completion_tokens || "?",
+        "completion-content": res?.choices[0]?.message?.content || "?",
+        "completion-function-call": JSON.stringify(res?.choices[0]?.message?.function_call || "{}"),
+      })
+
+      const content = res?.choices[0]?.message?.content
+      const functionCallResponse = res?.choices[0]?.message?.function_call
+      return {
+        content,
+        functionCall: functionCallResponse,
+        parsedArguments: content
+      };
+    })
+  }
+
+  private mapFunctionCallsToTools(functions: FunctionSpecification[], argCapture: (...args: any) => void): RunnableTools<any> {
+    return functions.map((functionSpec): { type: 'function', function: RunnableFunctionWithParse<any> } => {
+      return {
+        type: 'function',
+        function: {
+          name: functionSpec.name || "function_to_call",
+          function: argCapture,
+          description: functionSpec.description || "A function",
+          parse: (resp: string) => {
+            return functionSpec.parameters.parse(JSON.parse(resp))
+          },
+          parameters: zodToJsonSchema(functionSpec.parameters) as any, // TODO: why?!
+        },
+      }
+    })
   }
 }
