+++ core/tests/languageModels/functionlessLLM.spec.ts
@@ -1,9 +1,56 @@
 import { expect } from "chai";
-import { CortexStep, ChatMessageRoleEnum, decision, externalDialog, internalMonologue, questionMemory, z } from "../../src";
-import { FunctionlessLLM } from "../../src/languageModels/old_FunctionlessLLM";
+import { CortexStep, ChatMessageRoleEnum, decision, externalDialog, internalMonologue, questionMemory, z, brainstorm } from "../../src";
+import { FunctionlessLLM } from "../../src/languageModels/FunctionlessLLM";
 
 // this is set to skip because it requires a locally running LLM server.
-describe.only("FunctionlessLLM", () => {
+describe.skip("FunctionlessLLM", () => {
+  const step = new CortexStep("bob", {
+    processor: new FunctionlessLLM({
+      baseURL: "https://api.together.xyz/v1",
+      singleSystemMessage: true,
+      apiKey: process.env.TOGETHER_API_KEY,
+    }, {
+      // model: "mistralai/Mixtral-8x7B-Instruct-v0.1",
+      // model: "NousResearch/Nous-Hermes-2-Yi-34B",
+      model: "teknium/OpenHermes-2p5-Mistral-7B",
+      temperature: 0.7,
+      max_tokens: 300,
+    })
+  })
+
+  it('works with non-streaming, non-functions', async () => {
+    const result = await step.next(externalDialog("hi"))
+    expect(result.value).to.be.a("string")
+  })
+
+  it('works with streaming non-functions', async () => {
+    const { nextStep, stream } = await step.next(externalDialog("hi"), { stream: true })
+    expect(stream).to.be.an("AsyncGenerator")
+    let streamed = ""
+    for await (const res of stream) {
+      streamed += res
+    }
+    expect((await nextStep).value).to.be.a("string")
+    expect(streamed).to.equal((await nextStep).value)
+  })
+
+  it('streams functions', async () => {
+    const { nextStep, stream } = await step.next(brainstorm("hi"), { stream: true })
+
+    let streamed = ""
+    for await (const res of stream) {
+      streamed += res
+    }
+    expect((await nextStep).value).to.be.an("array")
+    expect(JSON.parse(streamed).new_ideas[0]).to.equal((await nextStep).value[0])
+  })
+
+  it("works with functions", async () => {
+    const result = await step.next(brainstorm("numbers less than 5"))
+    expect(result.value).to.be.an("array")
+    expect(parseInt(result.value[0])).to.be.a("number")
+  })
+
   it("runs example from readme", async () => {
     const queryMemory = (query: string) => {
       return () => {
@@ -17,16 +64,11 @@ describe.only("FunctionlessLLM", () => {
       }
     }
 
-    let step = new CortexStep("Jonathan", {
-      processor: new FunctionlessLLM({
-        baseURL: "http://localhost:1234/v1"
-      })
-    });
-    step = step.withMemory([{
+    const step2 = step.withMemory([{
       role: ChatMessageRoleEnum.System,
       content: "The name you are looking for is Jonathan"
     }])
-    const resp = await step.next(queryMemory("What is the name I'm looking for? Answer in a single word"))
+    const resp = await step2.next(queryMemory("What is the name I'm looking for? Answer in a single word"))
     expect(resp.value.answer).to.equal("Jonathan")
   }).timeout(65_000)
 
@@ -43,17 +85,11 @@ describe.only("FunctionlessLLM", () => {
       }
     }
 
-    let step = new CortexStep("Jonathan", {
-      processor: new FunctionlessLLM({
-        baseURL: "http://localhost:1234/v1",
-        singleSystemMessage: true,
-      })
-    });
-    step = step.withMemory([{
+    const step2 = step.withMemory([{
       role: ChatMessageRoleEnum.System,
       content: "The name you are looking for is Jonathan"
     }])
-    const resp = await step.next(queryMemory("What is the name I'm looking for? Answer in a single word"))
+    const resp = await step2.next(queryMemory("What is the name I'm looking for? Answer in a single word"))
     expect(resp.value.answer).to.equal("Jonathan")
   }).timeout(65_000)
 
@@ -73,8 +109,15 @@ describe.only("FunctionlessLLM", () => {
       ];
       const monologue = new CortexStep("Bogus", {
         processor: new FunctionlessLLM({
-          baseURL: "http://localhost:1234/v1",
+          baseURL: "https://api.together.xyz/v1",
           singleSystemMessage: true,
+          apiKey: process.env.TOGETHER_API_KEY,
+        }, {
+          // model: "mistralai/Mixtral-8x7B-Instruct-v0.1",
+          // model: "NousResearch/Nous-Hermes-2-Yi-34B",
+          model: "teknium/OpenHermes-2p5-Mistral-7B",
+          temperature: 0.7,
+          max_tokens: 300,
         })
       }).withMemory(memory)
